+++ crates/cubecl-matmul/src/components/config.rs
@@ -5,8 +5,7 @@ use std::hash::Hash;
 
 use crate::kernels::MatmulAvailabilityError;
 
-use super::problem::MatmulLineSizes;
-use super::{MatmulPrecision, MatmulProblem};
+use super::MatmulPrecision;
 
 pub type InvalidConfigError = Box<dyn Display>;
 
@@ -31,10 +30,8 @@ impl Display for FormattedConfigError {
 }
 
 /// Provides configuration for a matmul kernel at any level
-pub trait MatmulConfigFactory: Send + Sync + 'static {
-    /// Configuration tailored to the matmul implementation
+pub trait MatmulChecker: Send + Sync + 'static {
     type Config: MatmulConfig;
-    type Input;
 
     /// Asserts that the configuration for this matmul will lead to a valid computation
     fn check_config(config: &Self::Config) -> Result<(), InvalidConfigError>;
@@ -42,19 +39,9 @@ pub trait MatmulConfigFactory: Send + Sync + 'static {
     /// Checks if the client can handle the features used in this computation
     #[allow(clippy::result_large_err)]
     fn check_availability<R: Runtime, MP: MatmulPrecision>(
-        _client: &ComputeClient<R::Server, R::Channel>,
-        _config: &Self::Config,
+        client: &ComputeClient<R::Server, R::Channel>,
+        config: &Self::Config,
     ) -> Result<(), MatmulAvailabilityError>;
-
-    /// Create config for this matmul, given launch information
-    fn make_config(
-        input: Self::Input,
-        problem: &MatmulProblem,
-        line_sizes: &MatmulLineSizes,
-        cube_dim: &CubeDim,
-        cube_count: &CubeCount,
-        quantized: bool,
-    ) -> Self::Config;
 }
 
 /// A config for a matmul
