+++ crates/cubecl-matmul/src/components/global/single_stage/simple/matmul.rs
@@ -19,7 +19,7 @@ use std::marker::PhantomData;
 
 use crate::{
     components::{
-        Ident, InvalidConfigError, MatmulConfigFactory, MatmulProblem,
+        Ident, InvalidConfigError, MatmulProblem,
         global::{GlobalConfig, GlobalMatmulFamily},
         stage::{self, FullReaderFamily},
     },
@@ -45,6 +45,9 @@ where
     type Matmul<MP: MatmulPrecision> =
         SimpleMatmul<MP, SMM::Matmul<MP, LL::TilingLayout, RL::TilingLayout>, LL, RL>;
 
+    type Config = Config<SMM::Config>;
+    type Input = GlobalInput<SMM::Input>;
+
     fn cube_dim(
         selection: &MatmulSelection,
         load_specialization: LoadSpecializationConfig,
@@ -61,31 +64,8 @@ where
             ))
         }
     }
-}
 
-impl<SMM, LL, RL> MatmulConfigFactory for SimpleMatmulFamily<SMM, LL, RL>
-where
-    SMM: stage::StageMatmulFamily,
-    LL: SyncFullLoadingStrategy,
-    RL: SyncFullLoadingStrategy,
-{
-    type Input = GlobalInput<SMM::Input>;
-    type Config = Config<SMM::Config>;
-
-    fn check_config(config: &Self::Config) -> Result<(), InvalidConfigError> {
-        LL::check(config, Ident::Lhs)?;
-        RL::check(config, Ident::Rhs)?;
-        SMM::check_config(&config.stage_config())
-    }
-
-    fn check_availability<R: Runtime, MP: MatmulPrecision>(
-        client: &ComputeClient<R::Server, R::Channel>,
-        config: &Self::Config,
-    ) -> Result<(), MatmulAvailabilityError> {
-        SMM::check_availability::<R, MP>(client, &config.stage_config())
-    }
-
-    fn make_config(
+    fn setup(
         input: Self::Input,
         problem: &MatmulProblem,
         line_sizes: &MatmulLineSizes,
@@ -93,7 +73,7 @@ where
         cube_count: &CubeCount,
         quantized: bool,
     ) -> Self::Config {
-        let stage_config = SMM::make_config(
+        let stage_config = SMM::setup(
             input.stage_input,
             problem,
             line_sizes,
@@ -121,6 +101,29 @@ where
         )
     }
 }
+
+// impl<SMM, LL, RL> MatmulChecker for SimpleMatmulFamily<SMM, LL, RL>
+// where
+//     SMM: stage::StageMatmulFamily,
+//     LL: SyncFullLoadingStrategy,
+//     RL: SyncFullLoadingStrategy,
+// {
+//     type Input = GlobalInput<SMM::Input>;
+//     type Config = Config<SMM::Config>;
+
+//     fn check_config(config: &Self::Config) -> Result<(), InvalidConfigError> {
+//         LL::check(config, Ident::Lhs)?;
+//         RL::check(config, Ident::Rhs)?;
+//         SMM::check_config(&config.stage_config())
+//     }
+
+//     fn check_availability<R: Runtime, MP: MatmulPrecision>(
+//         client: &ComputeClient<R::Server, R::Channel>,
+//         config: &Self::Config,
+//     ) -> Result<(), MatmulAvailabilityError> {
+//         SMM::check_availability::<R, MP>(client, &config.stage_config())
+//     }
+// }
 
 /// Performs matrix multiplication at the global level, with each plane sharing the same responsibilities
 /// - All planes load data to the stage
