+++ crates/cubecl-matmul/src/components/stage/matmul/mod.rs
@@ -1,10 +1,5 @@
-mod partition_matmul;
-mod partitioned_stage_matmul;
+mod config;
+mod partition;
+mod partitioned_stage;
 
-pub mod plane;
-pub(super) mod shared;
-pub mod unit;
-
-pub use plane as plane_matmul;
-pub use shared::*;
-pub use unit as unit_matmul;
+pub use config::*;
