+++ crates/cubecl-matmul/src/components/stage/matmul/partition/mod.rs
@@ -1,0 +1,5 @@
+mod fragments;
+mod matmul;
+
+pub use fragments::*;
+pub use matmul::*;
