+++ crates/cubecl-matmul/src/components/stage/matmul/partitioned_stage/plane_partitioned/mod.rs
@@ -1,0 +1,4 @@
+mod matmul;
+mod setup;
+
+pub use matmul::*;
