+++ crates/cubecl-matmul/src/components/stage/matmul/partitioned_stage/plane_partitioned/setup.rs
@@ -1,53 +1,22 @@
 use crate::components::ComputeResources;
+use crate::components::MatmulChecker;
 use crate::components::MatmulProblem;
 use crate::components::TilingScheme;
 use crate::components::global::PlaneRoleConfig;
-use crate::components::global::PlaneWriter;
-use crate::components::global::RoleRule;
 use crate::components::stage::PartitionBuffering;
 use crate::components::stage::ReaderFamily;
-use crate::components::stage::shared::CommonStageConfig;
+use crate::components::stage::matmul::config::CommonStageConfig;
+use crate::components::stage::matmul::partitioned_stage::plane_partitioned::PlaneMatmul;
 use crate::components::stage::{StageConfig, StageMatmulFamily, TilingLayout};
 use crate::components::tile::TileConfig;
-use crate::components::tile::TileMatmulConfigInput;
 use crate::components::tile::TileMatmulFamily;
-use crate::components::{
-    InvalidConfigError, MatmulConfigFactory, MatmulLineSizes, MatmulPrecision,
-};
+use crate::components::tile::TileSetupInput;
+use crate::components::{InvalidConfigError, MatmulLineSizes, MatmulPrecision};
 use crate::kernels::MatmulAvailabilityError;
 use crate::kernels::matmul::StageInput;
 use core::marker::PhantomData;
 use cubecl::prelude::*;
 use cubecl_core as cubecl;
-use cubecl_std::tensor::r#virtual::{ReadWrite, VirtualTensor};
-
-use super::partitioned_stage_matmul::{PartitionedStageMatmul, StagePartitioner};
-
-type PlaneMatmul<MP, TMM, RL, RR> = PartitionedStageMatmul<MP, TMM, RL, RR, PlanePartitioner>;
-
-pub struct PlanePartitioner {}
-
-#[cube]
-impl StagePartitioner for PlanePartitioner {
-    type Writer<EO: Numeric> = PlaneWriter<EO>;
-
-    fn init_writer<EO: Numeric>(
-        tensor: VirtualTensor<EO, ReadWrite>,
-        x_offset: u32,
-        y_offset: u32,
-        batch_offset: u32,
-    ) -> Self::Writer<EO> {
-        PlaneWriter::<EO>::new(tensor, x_offset, y_offset, batch_offset)
-    }
-
-    fn position<S: StageConfig>(#[comptime] config: S) -> u32 {
-        RoleRule::new(config.role_rule_config()).compute_index()
-    }
-
-    fn num_primitives<S: StageConfig>(#[comptime] config: S) -> comptime_type!(u32) {
-        config.num_main_flow_planes()
-    }
-}
 
 pub struct PlaneMatmulFamily<TMM: TileMatmulFamily, LRF: ReaderFamily, RRF: ReaderFamily> {
     _phantom: PhantomData<(TMM, LRF, RRF)>,
@@ -61,6 +30,9 @@ impl<TMM: TileMatmulFamily, LRF: ReaderFamily, RRF: ReaderFamily> StageMatmulFam
     type Matmul<MP: MatmulPrecision, TL: TilingLayout, TR: TilingLayout> =
         PlaneMatmul<MP, TMM::Matmul<MP>, LRF::Reader<MP::ES, TL>, RRF::Reader<MP::ES, TR>>;
 
+    type Input = StageInput;
+    type Config = CommonStageConfig<TMM::Config>;
+
     fn computation_resources(
         tiling_scheme: &TilingScheme,
     ) -> Result<ComputeResources, InvalidConfigError> {
@@ -72,13 +44,47 @@ impl<TMM: TileMatmulFamily, LRF: ReaderFamily, RRF: ReaderFamily> StageMatmulFam
             unreachable!("Plane matmul should not demand units")
         }
     }
+
+    fn setup(
+        stage_input: Self::Input,
+        problem: &MatmulProblem,
+        line_sizes: &MatmulLineSizes,
+        cube_dim: &CubeDim,
+        quantized: bool,
+    ) -> Self::Config {
+        let tile_input = TileSetupInput {
+            vectorization: stage_input.stage_vectorization,
+            tile_size: stage_input.tiling_scheme.tile_size,
+        };
+        let tile_config = TMM::setup(tile_input, problem, line_sizes, cube_dim);
+
+        let compute_planes =
+            <Self as StageMatmulFamily>::computation_resources(&stage_input.tiling_scheme)
+                .unwrap_or_else(|e| panic!("{}", e))
+                .as_plane_resources(tile_config.plane_dim())
+                .unwrap_or_else(|e| panic!("{}", e))
+                .get_count();
+        let plane_role_config = PlaneRoleConfig::from_plane_roles(
+            stage_input
+                .load_specialization
+                .to_plane_roles(compute_planes),
+        );
+
+        CommonStageConfig::new(
+            tile_config,
+            stage_input.tiling_scheme,
+            quantized,
+            stage_input.partition_buffering,
+            stage_input.num_stages,
+            plane_role_config,
+        )
+    }
 }
 
-impl<TMM: TileMatmulFamily, LRF: ReaderFamily, RRF: ReaderFamily> MatmulConfigFactory
+impl<TMM: TileMatmulFamily, LRF: ReaderFamily, RRF: ReaderFamily> MatmulChecker
     for PlaneMatmulFamily<TMM, LRF, RRF>
 {
-    type Input = StageInput;
-    type Config = CommonStageConfig<TMM::Config>;
+    type Config = CommonStageConfig<<TMM as MatmulChecker>::Config>;
 
     fn check_config(config: &Self::Config) -> Result<(), InvalidConfigError> {
         let num_planes_needed = config.tiling_scheme().stage_partitions_in_stage_mn();
@@ -107,43 +113,5 @@ impl<TMM: TileMatmulFamily, LRF: ReaderFamily, RRF: ReaderFamily> MatmulConfigFa
         config: &Self::Config,
     ) -> Result<(), MatmulAvailabilityError> {
         TMM::check_availability::<R, MP>(client, &config.tile_config)
-    }
-
-    fn make_config(
-        stage_input: Self::Input,
-        problem: &MatmulProblem,
-        line_sizes: &MatmulLineSizes,
-        cube_dim: &CubeDim,
-        cube_count: &CubeCount,
-        quantized: bool,
-    ) -> Self::Config {
-        let tile_input = TileMatmulConfigInput {
-            vectorization: stage_input.stage_vectorization,
-            tile_size: stage_input.tiling_scheme.tile_size,
-        };
-        let tile_config = TMM::make_config(
-            tile_input, problem, line_sizes, cube_dim, cube_count, quantized,
-        );
-
-        let compute_planes =
-            <Self as StageMatmulFamily>::computation_resources(&stage_input.tiling_scheme)
-                .unwrap_or_else(|e| panic!("{}", e))
-                .as_plane_resources(tile_config.plane_dim())
-                .unwrap_or_else(|e| panic!("{}", e))
-                .get_count();
-        let plane_role_config = PlaneRoleConfig::from_plane_roles(
-            stage_input
-                .load_specialization
-                .to_plane_roles(compute_planes),
-        );
-
-        CommonStageConfig::new(
-            tile_config,
-            stage_input.tiling_scheme,
-            quantized,
-            stage_input.partition_buffering,
-            stage_input.num_stages,
-            plane_role_config,
-        )
     }
 }
