+++ crates/cubecl-matmul/src/components/tile/accelerated/mod.rs
@@ -1,0 +1,3 @@
+mod config;
+mod matmul;
+mod setup;
